fix bug when special xml escaped characters encountered
the saxparser will sometimes break a string of text between tags such
as <author>Joachim M&uuml;ller</author>
into multiple callback calls to the character() callback.  we need to
handle this by creating a StringBuilder and appending the multiple
string subsections together and then adding the final string when we
see the end tag.